smallbaselineApp: check key 'mintpy.networkInversion.shadowMask' (#218)
* smallbaselineApp: check key 'mintpy.networkInversion.shadowMask'

smallbaselineApp: check for 'mintpy.networkInversion.shadowMask' key for update mode while generating the mask from temporal coherence.

asc_desc2horz_vert: constrain the output data type to float32

utils/plot: add buffer_year argument in auto_adjust_xaxis_date()

docs/google_earth.md: remove commands example as the user should check the --help from the script, which will be the most up-to-date.

sh/plot_smallbaselineApp.sh: remove the trop_models variable definition and use for loop directly.

+++ mintpy/asc_desc2horz_vert.py
@@ -117,7 +117,7 @@ def main(iargs=None):
     length = int(round((south - north) / lat_step))
 
     # Read data in common AOI: LOS displacement, heading angle, incident angle
-    u_los = np.zeros((2, width*length))
+    u_los = np.zeros((2, width*length), dtype=np.float32)
     heading = []
     incidence = []
     for i in range(len(inps.file)):
@@ -156,7 +156,7 @@ def main(iargs=None):
         A[i, 1] = np.sin(incidence[i]) * np.sin(heading[i]-inps.azimuth)
 
     A_inv = np.linalg.pinv(A)
-    u_vh = np.dot(A_inv, u_los)
+    u_vh = np.dot(A_inv, u_los).astype(np.float32)
 
     u_v = np.reshape(u_vh[0, :], (length, width))
     u_h = np.reshape(u_vh[1, :], (length, width))
